+++ nextgisweb/auth/__init__.py
@@ -172,6 +172,7 @@ class AuthComponent(Component):
                     self.oauth.options['server.base_url']
                     if (enabled and 'server.base_url' in self.oauth.options)
                     else None),
+                group_mapping='profile.member_of.attr' in self.oauth.options if enabled else False,
             )
         )