+++ nextgisweb/auth/migration/39b20e77-group-oauth_mapping-column.rw.sql
@@ -1,0 +1,3 @@
+/*** { "revision": "39b20e77" } ***/
+
+ALTER TABLE auth_group DROP COLUMN oauth_mapping;